Create News Type Classifiers using various Models
This is the code that was used to analyze the accuracy of various types of classification algorithms in predicting the type of news reports. The dataset used is fetch_20newsgroups from sklearn and the models of classifier used are Multinomial Naive Bayes, KNN and Random Forest Classifier. The function used for "vectorizing" the news articles was TfidfVectorizer from sklearn.